+++ my_iot/services/clock.py
@@ -7,8 +7,8 @@ from typing import Optional
 
 from loguru import logger
 
-from myiot.services.base import Service
-from myiot.types_ import Event
+from my_iot.services.base import Service
+from my_iot.types_ import Event
 
 
 class Clock(Service):