+++ my_iot/web.py
@@ -9,16 +9,16 @@ from aiohttp_jinja2 import template
 from loguru import logger
 from pkg_resources import resource_string
 
-from myiot import templates
-from myiot.automation import Automation
-from myiot.constants import ACTUAL_KEY
-from myiot.context import Context
-from myiot.decorators import authenticate_user
-from myiot.types_ import Event
+from my_iot import templates
+from my_iot.automation import Automation
+from my_iot.constants import ACTUAL_KEY
+from my_iot.context import Context
+from my_iot.decorators import authenticate_user
+from my_iot.types_ import Event
 
 routes = web.RouteTableDef()
 statics = {
-    name: resource_string('myiot', f'static/{name}')
+    name: resource_string('my_iot', f'static/{name}')
     for name in [
         'android-chrome-192x192.png',
         'android-chrome-512x512.png',